Updated production build for release
