+++ crawlerapp/public/assets/frontend-22dc4f1338193f64ee47192b60bd0619.css
@@ -1,0 +1,18 @@
+@import 'components/user-input-form-c6782d90c46df6a21e410e06e2f70e48.css';
+@import 'components/nav-bar-f4cde15da5f63207be826fbfb8ac51f6.css';
+
+.lower-components {
+    display: flex;
+}
+
+.history-component {
+    flex-basis:  40%;
+}
+
+.backend-list {
+    margin-left: 300px; 
+}
+
+.loading-spinner {
+    margin: 5% 0 0 25%;
+}
