+++ crawlerapp/public/assets/frontend-5e1acd615063eb95f0cd189ca5427bf2.js
@@ -1,0 +1,21 @@
+"use strict"
+define("frontend/adapters/application",["exports","ember-local-storage/adapters/local"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("frontend/app",["exports","frontend/resolver","ember-load-initializers","frontend/config/environment"],function(e,t,n,r){Object.defineProperty(e,"__esModule",{value:!0})
+var a=Ember.Application.extend({modulePrefix:r.default.modulePrefix,podModulePrefix:r.default.podModulePrefix,Resolver:t.default});(0,n.default)(a,r.default.modulePrefix),e.default=a}),define("frontend/components/nav-bar",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.Component.extend({})}),define("frontend/components/radio-button-input",["exports","ember-radio-button/components/radio-button-input"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("frontend/components/radio-button",["exports","ember-radio-button/components/radio-button"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("frontend/components/user-input-form",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.Component.extend({store:Ember.inject.service(),router:Ember.inject.service(),searchMethod:"bfs",actions:{submitForm:function(){var e=this.store.createRecord("search-request",{url:this.get("url"),searchMethod:this.searchMethod,depth:this.get("depth"),keyword:this.get("keyword")})
+e.save(),console.log(e.id),this.set("url",""),this.set("searchMethod","bfs"),this.set("depth",""),this.set("keyword",""),this.get("router").transitionTo("/dashboard/search/"+e.id)},methodChanged:function(e){this.set("searchMethod",e)}}})}),define("frontend/controllers/dashboard",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.Controller.extend({actions:{transitionInController:function(){this.transitionToRoute("dashboard")},deleteSearch:function(e){this.store.findRecord("search-request",e).then(function(e){return e.destroyRecord()})},repeatSearch:function(e){this.transitionToRoute("/dashboard/search/"+e)}}})}),define("frontend/helpers/app-version",["exports","frontend/config/environment","ember-cli-app-version/utils/regexp"],function(e,t,n){function r(e){var r=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},a=t.default.APP.version,o=r.versionOnly||r.hideSha,i=r.shaOnly||r.hideVersion,l=null
+return o&&(r.showExtended&&(l=a.match(n.versionExtendedRegExp)),l||(l=a.match(n.versionRegExp))),i&&(l=a.match(n.shaRegExp)),l?l[0]:a}Object.defineProperty(e,"__esModule",{value:!0}),e.appVersion=r,e.default=Ember.Helper.helper(r)}),define("frontend/helpers/pluralize",["exports","ember-inflector/lib/helpers/pluralize"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default=t.default}),define("frontend/helpers/singularize",["exports","ember-inflector/lib/helpers/singularize"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default=t.default}),define("frontend/initializers/app-version",["exports","ember-cli-app-version/initializer-factory","frontend/config/environment"],function(e,t,n){Object.defineProperty(e,"__esModule",{value:!0})
+var r=void 0,a=void 0
+n.default.APP&&(r=n.default.APP.name,a=n.default.APP.version),e.default={name:"App Version",initialize:(0,t.default)(r,a)}}),define("frontend/initializers/container-debug-adapter",["exports","ember-resolver/resolvers/classic/container-debug-adapter"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default={name:"container-debug-adapter",initialize:function(){var e=arguments[1]||arguments[0]
+e.register("container-debug-adapter:main",t.default),e.inject("container-debug-adapter:main","namespace","application:main")}}}),define("frontend/initializers/ember-data",["exports","ember-data/setup-container","ember-data"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default={name:"ember-data",initialize:t.default}}),define("frontend/initializers/export-application-global",["exports","frontend/config/environment"],function(e,t){function n(){var e=arguments[1]||arguments[0]
+if(!1!==t.default.exportApplicationGlobal){var n
+if("undefined"!=typeof window)n=window
+else if("undefined"!=typeof global)n=global
+else{if("undefined"==typeof self)return
+n=self}var r,a=t.default.exportApplicationGlobal
+r="string"==typeof a?a:Ember.String.classify(t.default.modulePrefix),n[r]||(n[r]=e,e.reopen({willDestroy:function(){this._super.apply(this,arguments),delete n[r]}}))}}Object.defineProperty(e,"__esModule",{value:!0}),e.initialize=n,e.default={name:"export-application-global",initialize:n}}),define("frontend/initializers/local-storage-adapter",["exports","ember-local-storage/initializers/local-storage-adapter"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}}),Object.defineProperty(e,"initialize",{enumerable:!0,get:function(){return t.initialize}})}),define("frontend/instance-initializers/ember-data",["exports","ember-data/initialize-store-service"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default={name:"ember-data",initialize:t.default}}),define("frontend/models/search-request",["exports","ember-data"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default=t.default.Model.extend({url:t.default.attr("string"),searchMethod:t.default.attr("string"),depth:t.default.attr("number"),keyword:t.default.attr("string")})}),define("frontend/resolver",["exports","ember-resolver"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default=t.default}),define("frontend/router",["exports","frontend/config/environment"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0})
+var n=Ember.Router.extend({location:t.default.locationType,rootURL:t.default.rootURL})
+n.map(function(){this.route("dashboard",function(){this.route("search",{path:"/search/:id"})})}),e.default=n}),define("frontend/routes/dashboard",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.Route.extend({model:function(){return this.store.findAll("search-request")}})}),define("frontend/routes/dashboard/search",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.Route.extend({beforeModel:function(){Ember.$(".show-loading").show()},model:function(e){var t=this.store.findRecord("search-request",e.id),n=t.then(function(e){return e.data}).then(function(e){return function(e){console.log("formdata apicall",e)
+var t="/api/search/?url="+e.url+"&searchMethod="+e.searchMethod+"&depth="+e.depth+"&keyword="+e.keyword
+return fetch(t).then(function(e){return e.json()}).then(function(e){return console.log(e),e})}(e)})
+return Ember.RSVP.hash({data:n})},afterModel:function(){Ember.$(".show-loading").hide()}})}),define("frontend/routes/index",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.Route.extend({beforeModel:function(){this.transitionTo("dashboard")}})}),define("frontend/serializers/application",["exports","ember-local-storage/serializers/serializer"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("frontend/services/ajax",["exports","ember-ajax/services/ajax"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("frontend/templates/application",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.HTMLBars.template({id:"YnWrGo4X",block:'{"symbols":[],"statements":[[1,[21,"nav-bar"],false],[0,"\\n\\n"],[1,[21,"outlet"],false]],"hasEval":false}',meta:{moduleName:"frontend/templates/application.hbs"}})}),define("frontend/templates/components/nav-bar",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.HTMLBars.template({id:"hZDdZytm",block:'{"symbols":[],"statements":[[7,"nav"],[9],[0,"\\n    "],[7,"div"],[11,"class","nav-wrapper"],[9],[0,"\\n      "],[7,"a"],[11,"href","/"],[11,"class","brand-logo center"],[9],[0,"Team Azha Web Crawler"],[10],[0,"\\n      "],[7,"ul"],[11,"id","nav-mobile"],[11,"class","right"],[9],[0,"\\n        "],[7,"li"],[9],[7,"a"],[11,"href","/"],[9],[0,"About"],[10],[10],[0,"\\n      "],[10],[0,"\\n    "],[10],[0,"\\n"],[10],[0,"\\n"]],"hasEval":false}',meta:{moduleName:"frontend/templates/components/nav-bar.hbs"}})}),define("frontend/templates/components/user-input-form",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.HTMLBars.template({id:"mVJ0P/Ht",block:'{"symbols":[],"statements":[[7,"div"],[11,"class","form-container"],[9],[0,"\\n"],[7,"form"],[11,"class","input-form"],[3,"action",[[22,0,[]],"submitForm"],[["on"],["submit"]]],[9],[0,"\\n    "],[7,"div"],[11,"class","input-field col s6"],[9],[0,"\\n        "],[1,[27,"input",null,[["value","id","type"],[[23,["url"]],"search-url","text"]]],false],[0,"\\n        "],[7,"label"],[11,"for","search-url"],[9],[0,"Url to Search"],[10],[0,"\\n    "],[10],[0,"\\n    "],[7,"p"],[9],[0,"\\n"],[4,"radio-button",null,[["value","groupValue","changed"],["bfs",[23,["searchMethod"]],[27,"action",[[22,0,[]],"methodChanged"],null]]],{"statements":[[0,"        "],[7,"span"],[9],[0,"Breadth First Search"],[10],[0,"\\n"]],"parameters":[]},null],[0,"    "],[10],[0,"\\n    "],[7,"p"],[9],[0,"\\n"],[4,"radio-button",null,[["value","groupValue","changed"],["dfs",[23,["searchMethod"]],[27,"action",[[22,0,[]],"methodChanged"],null]]],{"statements":[[0,"        "],[7,"span"],[9],[0,"Depth First Search"],[10],[0,"\\n"]],"parameters":[]},null],[0,"    "],[10],[0,"\\n    "],[7,"div"],[11,"class","input-field col s6"],[9],[0,"\\n        "],[1,[27,"input",null,[["value","id","type"],[[23,["depth"]],"search-depth","text"]]],false],[0,"\\n        "],[7,"label"],[11,"for","search-depth"],[9],[0,"Search Depth"],[10],[0,"\\n    "],[10],[0,"\\n        "],[7,"div"],[11,"class","input-field col s6"],[9],[0,"\\n        "],[1,[27,"input",null,[["value","id","type"],[[23,["keyword"]],"search-keyword","text"]]],false],[0,"\\n        "],[7,"label"],[11,"for","search-keyword"],[9],[0,"Keyword"],[10],[0,"\\n    "],[10],[0,"\\n     "],[7,"button"],[11,"class","btn waves-effect waves-light"],[11,"type","submit"],[9],[0,"Search\\n        "],[7,"i"],[11,"class","material-icons right"],[9],[0,"search"],[10],[0,"\\n    "],[10],[0,"\\n"],[10],[0,"  \\n"],[10]],"hasEval":false}',meta:{moduleName:"frontend/templates/components/user-input-form.hbs"}})}),define("frontend/templates/dashboard",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.HTMLBars.template({id:"mZhjHAfm",block:'{"symbols":["request"],"statements":[[1,[27,"user-input-form",null,[["model","transitionInComponent"],[[23,["model"]],[27,"action",[[22,0,[]],"transitionInController"],null]]]],false],[0,"\\n\\n"],[7,"div"],[11,"class","lower-components"],[9],[0,"\\n    "],[7,"div"],[11,"class","history-component"],[9],[0,"\\n        "],[7,"table"],[9],[0,"\\n            "],[7,"tr"],[9],[0,"\\n                "],[7,"th"],[9],[0,"Id"],[10],[0,"\\n                "],[7,"th"],[9],[0,"URL"],[10],[0,"\\n                "],[7,"th"],[9],[0,"Search Method"],[10],[0,"\\n                "],[7,"th"],[9],[0,"Depth"],[10],[0,"\\n                "],[7,"th"],[9],[0,"Keyword"],[10],[0,"\\n                "],[7,"th"],[9],[0,"Repeat Search"],[10],[0,"\\n                "],[7,"th"],[9],[0,"Delete"],[10],[0,"\\n            "],[10],[0,"\\n"],[4,"each",[[23,["model"]]],null,{"statements":[[0,"            "],[7,"tr"],[9],[0,"\\n                "],[7,"td"],[9],[1,[22,1,["id"]],false],[10],[0,"\\n                "],[7,"td"],[9],[1,[22,1,["url"]],false],[10],[0,"\\n                "],[7,"td"],[9],[1,[22,1,["searchMethod"]],false],[10],[0,"\\n                "],[7,"td"],[9],[1,[22,1,["depth"]],false],[10],[0,"\\n                "],[7,"td"],[9],[1,[22,1,["keyword"]],false],[10],[0,"\\n                "],[7,"td"],[9],[0,"\\n                    "],[7,"a"],[11,"class","btn-floating btn-medium waves-effect waves-light blue"],[3,"action",[[22,0,[]],"repeatSearch",[22,1,["id"]]]],[9],[7,"i"],[11,"class","material-icons"],[9],[0,"youtube_searched_for"],[10],[10],[0,"\\n                "],[10],[0,"\\n                "],[7,"td"],[9],[0,"\\n                    "],[7,"a"],[11,"class","btn-floating btn-medium waves-effect waves-light red"],[3,"action",[[22,0,[]],"deleteSearch",[22,1,["id"]]]],[9],[7,"i"],[11,"class","material-icons"],[9],[0,"clear"],[10],[10],[0,"\\n                "],[10],[0,"\\n            "],[10],[0,"\\n"]],"parameters":[1]},null],[0,"        "],[10],[0,"\\n    "],[10],[0,"\\n\\n\\n"],[7,"div"],[11,"class","show-loading loading-spinner"],[11,"style","display: none;"],[9],[0,"\\n      "],[7,"div"],[11,"class","preloader-wrapper big active"],[9],[0,"\\n    "],[7,"div"],[11,"class","spinner-layer spinner-blue-only"],[9],[0,"\\n      "],[7,"div"],[11,"class","circle-clipper left"],[9],[0,"\\n        "],[7,"div"],[11,"class","circle"],[9],[10],[0,"\\n      "],[10],[7,"div"],[11,"class","gap-patch"],[9],[0,"\\n        "],[7,"div"],[11,"class","circle"],[9],[10],[0,"\\n      "],[10],[7,"div"],[11,"class","circle-clipper right"],[9],[0,"\\n        "],[7,"div"],[11,"class","circle"],[9],[10],[0,"\\n      "],[10],[0,"\\n    "],[10],[0,"\\n  "],[10],[0,"\\n"],[10],[0,"\\n"],[1,[21,"outlet"],false],[0,"\\n\\n"],[10],[0,"\\n"]],"hasEval":false}',meta:{moduleName:"frontend/templates/dashboard.hbs"}})}),define("frontend/templates/dashboard/search",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.HTMLBars.template({id:"SJ87dMzK",block:'{"symbols":["other","links"],"statements":[[0,"\\n"],[7,"div"],[11,"class","backend-list"],[9],[0,"\\n    "],[7,"ul"],[9],[0," "],[7,"h3"],[9],[0,"Data Received from Backend"],[10],[0,"\\n"],[4,"each",[[27,"-each-in",[[23,["model","data","links"]]],null]],null,{"statements":[[0,"            "],[7,"li"],[9],[0,"\\n                "],[1,[22,2,[]],false],[0,"\\n            "],[10],[0,"\\n"]],"parameters":[1,2]},null],[0,"    "],[10],[0,"\\n"],[10],[0,"\\n"]],"hasEval":false}',meta:{moduleName:"frontend/templates/dashboard/search.hbs"}})}),define("frontend/templates/index",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.HTMLBars.template({id:"TOSW2YS6",block:'{"symbols":[],"statements":[[1,[21,"outlet"],false]],"hasEval":false}',meta:{moduleName:"frontend/templates/index.hbs"}})})
+define("frontend/config/environment",[],function(){try{var e="frontend/config/environment",t=document.querySelector('meta[name="'+e+'"]').getAttribute("content"),n={default:JSON.parse(unescape(t))}
+return Object.defineProperty(n,"__esModule",{value:!0}),n}catch(r){throw new Error('Could not read config from meta tag with name "'+e+'".')}}),runningTests||require("frontend/app").default.create({name:"frontend",version:"0.0.0+3a41457e"})
