+++ crawlerapp/public/index.html
@@ -8,7 +8,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1">
 
     
-<meta name="frontend/config/environment" content="%7B%22modulePrefix%22%3A%22frontend%22%2C%22environment%22%3A%22production%22%2C%22rootURL%22%3A%22/%22%2C%22locationType%22%3A%22auto%22%2C%22EmberENV%22%3A%7B%22FEATURES%22%3A%7B%7D%2C%22EXTEND_PROTOTYPES%22%3A%7B%22Date%22%3Afalse%7D%7D%2C%22APP%22%3A%7B%22name%22%3A%22frontend%22%2C%22version%22%3A%220.0.0+4ee8b2c5%22%7D%2C%22exportApplicationGlobal%22%3Afalse%7D" />
+<meta name="frontend/config/environment" content="%7B%22modulePrefix%22%3A%22frontend%22%2C%22environment%22%3A%22production%22%2C%22rootURL%22%3A%22/%22%2C%22locationType%22%3A%22auto%22%2C%22EmberENV%22%3A%7B%22FEATURES%22%3A%7B%7D%2C%22EXTEND_PROTOTYPES%22%3A%7B%22Date%22%3Afalse%7D%7D%2C%22APP%22%3A%7B%22name%22%3A%22frontend%22%2C%22version%22%3A%220.0.0+3a41457e%22%7D%2C%22exportApplicationGlobal%22%3Afalse%7D" />
     <!-- Compiled and minified CSS -->
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/css/materialize.min.css">
     
@@ -18,7 +18,7 @@
 
     <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
     <link integrity="" rel="stylesheet" href="/assets/vendor-d41d8cd98f00b204e9800998ecf8427e.css">
-    <link integrity="" rel="stylesheet" href="/assets/frontend-afd6b727b11e13bc767f0ebd17c47ecd.css">
+    <link integrity="" rel="stylesheet" href="/assets/frontend-22dc4f1338193f64ee47192b60bd0619.css">
 
     
   </head>
@@ -26,7 +26,7 @@
     
 
     <script src="/assets/vendor-5bec10d902faf06565f5cc7b5fd59831.js" integrity="sha256-TnpXdN91gQjexjowM0kWvDtLVl/pE3RaeiiTG+CpuDE= sha512-D3Lu1Kq95RQkc3wUx1beq0wY/SSJ8JMCAvMIViJDH4pQ5r+b+d3V/ewUw8Vaev1htyoHf9fn0Us4K1S0OTk00g==" ></script>
-    <script src="/assets/frontend-c31ff617789324c6e42eef934420851b.js" integrity="sha256-Gs/Jb76DHEpH3xDG3pIa6salIUlAygvx3hfANnSJP/Y= sha512-/Nj96smCvtp6XhcfKDx42CHRjF+n49TezkLj2toEXIKT9uibYeESw+Og+rbNDgnAV9rfa3wI6xqBiozlHzIuUQ==" ></script>
+    <script src="/assets/frontend-5e1acd615063eb95f0cd189ca5427bf2.js" integrity="sha256-eWyi2X2aa8lfisEVNhMaVpTyR1hsDhMxz0EiVlxLlRc= sha512-C1qGjTUUO2btAZERVMK1EAbL6GzHxbhuyXOR4Uq1ghxwPdBQfU984lIPWRoVpcPoTwHS9zDcboVqYC6vQQj4Zg==" ></script>
 
     
   </body>
